+++ packages/common/src/content/_internal/computeProps.ts
@@ -1,11 +1,15 @@
 import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { UserSession } from "@esri/arcgis-rest-auth";
 import { getItemThumbnailUrl } from "../../resources";
-import { IHubEditableContent, IHubLocation } from "../../core";
 import { IModel } from "../../types";
 import { bBoxToExtent, extentToPolygon, isBBox } from "../../extent";
 import { IExtent } from "@esri/arcgis-rest-types";
 import Geometry = __esri.Geometry;
+import { getItemHomeUrl } from "../../urls/get-item-home-url";
+import { getHubRelativeUrl } from "./internalContentUtils";
+import { IHubLocation } from "../../core/types/IHubLocation";
+import { IHubEditableContent } from "../../core/types/IHubEditableContent";
+import { getRelativeWorkspaceUrl } from "../../core/getRelativeWorkspaceUrl";
 
 // if called and valid, set 3 things -- else just return type custom
 export const getItemExtent = (itemExtent: number[][]): IExtent => {
@@ -40,7 +44,20 @@ export function computeProps(
     token = session.token;
   }
   // thumbnail url
-  content.thumbnailUrl = getItemThumbnailUrl(model.item, requestOptions, token);
+  const thumbnailUrl = getItemThumbnailUrl(model.item, requestOptions, token);
+  // TODO: Remove this once opendata-ui starts using `links.thumbnail` instead
+  content.thumbnailUrl = thumbnailUrl;
+  content.links = {
+    self: getItemHomeUrl(content.id, requestOptions),
+    siteRelative: getHubRelativeUrl(
+      content.type,
+      content.slug || content.id,
+      content.typeKeywords
+    ),
+    workspaceRelative: getRelativeWorkspaceUrl("content", content.id),
+    thumbnail: thumbnailUrl,
+  };
+  content.licenseInfo = model.item.licenseInfo;
 
   if (!content.location) {
     // build location if one does not exist based off of the boundary and the item's extent