+++ packages/common/src/groups/add-users-workflow/index.ts
@@ -1,5 +1,4 @@
 export * from "./output-processors";
 export * from "./utils";
-export * from "./workflow-sections";
 export * from "./add-users-to-group";
 export * from "./interfaces";