+++ packages/common/src/groups/index.ts
@@ -3,3 +3,11 @@ export * from "./add-users-workflow";
 export * from "./types";
 export * from "./HubGroups";
 export * from "./HubGroup";
+export * from "./addOrInviteUsersToGroup";
+export * from "./addOrInviteUsersToGroups";
+// TODO: The below are being used in hub-teams. When we deprecate that package we can move
+// The below into _internal and remove the exports from here. They were previously in
+// the add-users-workflow directory
+export * from "./autoAddUsers";
+export * from "./inviteUsers";
+export * from "./emailOrgUsers";