+++ packages/common/src/groups/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./IGroupMembershipSummary";
+export * from "./types";