+++ packages/common/src/pages/_internal/computeProps.ts
@@ -1,11 +1,14 @@
 import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { UserSession } from "@esri/arcgis-rest-auth";
 import { getItemThumbnailUrl } from "../../resources";
-import { IHubPage } from "../../core";
 import { IModel } from "../../types";
 
 import { processEntityFeatures } from "../../permissions/_internal/processEntityFeatures";
 import { PageDefaultFeatures } from "./PageBusinessRules";
+import { getItemHomeUrl } from "../../urls/get-item-home-url";
+import { IHubPage } from "../../core/types/IHubPage";
+import { getHubRelativeUrl } from "../../content/_internal/internalContentUtils";
+import { getRelativeWorkspaceUrl } from "../../core/getRelativeWorkspaceUrl";
 
 /**
  * Given a model and a page, set various computed properties that can't be directly mapped
@@ -26,7 +29,16 @@ export function computeProps(
     token = session.token;
   }
   // thumbnail url
-  page.thumbnailUrl = getItemThumbnailUrl(model.item, requestOptions, token);
+  const thumbnailUrl = getItemThumbnailUrl(model.item, requestOptions, token);
+  // TODO: Remove this once opendata-ui starts using `links.thumbnail` instead
+  page.thumbnailUrl = thumbnailUrl;
+  page.links = {
+    self: getItemHomeUrl(page.id, requestOptions),
+    siteRelative: `/pages/${page.id}`,
+    workspaceRelative: getRelativeWorkspaceUrl("page", page.id),
+    layoutRelative: `/pages/${page.id}/edit`,
+    thumbnail: thumbnailUrl,
+  };
 
   // Handle Dates
   page.createdDate = new Date(model.item.created);