+++ packages/common/src/permissions/checkPermission.ts
@@ -2,16 +2,9 @@ import { IArcGISContext } from "../ArcGISContext";
 import { getWithDefault } from "../objects";
 import { checkLicense } from "./_internal/checkLicense";
 import { getPermissionPolicy } from "./HubPermissionPolicies";
-import {
-  IPermissionAccessResponse,
-  Permission,
-  isPermission,
-  IPolicyCheck,
-  IEntityPermissionPolicy,
-} from "./types";
+import { Permission, isPermission } from "./types/Permission";
 import { getPolicyResponseCode } from "./_internal/getPolicyResponseCode";
 import { checkAuthentication } from "./_internal/checkAuthentication";
-import { checkAlphaGating } from "./_internal/checkAlphaGating";
 import { checkOwner } from "./_internal/checkOwner";
 import { checkEdit } from "./_internal/checkEdit";
 import { checkPrivileges } from "./_internal/checkPrivileges";
@@ -20,8 +13,12 @@ import { checkAssertions } from "./_internal/checkAssertions";
 import { checkParents } from "./_internal/checkParents";
 import { checkEnvironment } from "./_internal/checkEnvironment";
 import { checkAvailability } from "./_internal/checkAvailability";
-import { checkEntityFeature } from "./_internal/checkEntityFeature";
 import { checkServiceStatus } from "./_internal/checkServiceStatus";
+import { PermissionCheckFunction } from "./_internal/PermissionCheckFunction";
+import { IPermissionAccessResponse } from "./types/IPermissionAccessResponse";
+import { PolicyResponse } from "./types/PolicyResponse";
+import { IPolicyCheck } from "./types/IPolicyCheck";
+import { IEntityPermissionPolicy } from "./types/IEntityPermissionPolicy";
 
 /**
  * Type to allow either an entity or and entity and label to be
@@ -50,105 +47,183 @@ export function checkPermission(
 ): IPermissionAccessResponse {
   const label = entityOrOptions?.label || "";
   const entity = entityOrOptions?.entity || entityOrOptions;
-  // Default to granted
-  let response: IPermissionAccessResponse = {
-    policy: permission,
-    access: true,
-    response: "granted",
-    code: getPolicyResponseCode("granted"),
-    checks: [],
-  };
+
   // Is this even a valid permission?
   if (!isPermission(permission)) {
-    response = {
+    const invalidPermissionResponse = {
       policy: permission,
       access: false,
       response: "invalid-permission",
       code: getPolicyResponseCode("invalid-permission"),
       checks: [],
     } as IPermissionAccessResponse;
-  } else {
-    // Get the system policy for this permission
-    const systemPolicy = getPermissionPolicy(permission);
+    logResponse(invalidPermissionResponse, label);
+    return invalidPermissionResponse;
+  }
 
-    // TODO: handle null systemPolicy
+  // Get the system policy for this permission
+  const systemPolicy = getPermissionPolicy(permission);
 
-    const checks = [
-      checkParents,
-      checkServiceStatus,
-      checkEntityFeature,
-      checkAuthentication,
-      checkEnvironment,
-      checkAvailability,
-      checkLicense,
-      checkPrivileges,
-      checkOwner,
-      checkEdit,
-      checkAssertions,
-      checkAlphaGating, // TODO: Remove with Capability Refactor
-    ].reduce((acc: IPolicyCheck[], fn) => {
-      acc = [...acc, ...fn(systemPolicy, context, entity)];
-      return acc;
-    }, []);
+  // handle null systemPolicy
+  if (!systemPolicy) {
+    const missingPolicyResponse = {
+      policy: permission,
+      access: false,
+      response: "no-policy-exists",
+      code: getPolicyResponseCode("no-policy-exists"),
+      checks: [],
+    } as IPermissionAccessResponse;
+    logResponse(missingPolicyResponse, label);
+    return missingPolicyResponse;
+  }
 
-    // For system policies, all conditions must be met, so we can
-    // iterate through the checks and set the response to the first failure
-    // while still returning all the checks for observability
-    checks.forEach((check) => {
-      if (check.response !== "granted" && response.response === "granted") {
-        response.response = check.response;
-        response.code = check.code;
-        response.access = false;
-      }
-    });
+  const flagging = {
+    hasFlag: false,
+    value: false,
+    type: "none",
+  };
 
-    response.checks = checks;
+  // Entity Feature Flags
+  // aka how we disable "features" on a per-entity basis
 
-    // Entity policies are treated as "grants" so we only need to pass one
-    if (entity) {
-      const entityPolicies: IEntityPermissionPolicy[] = getWithDefault(
-        entity,
-        "permissions",
-        []
-      );
-
-      const entityPermissionPolicies = entityPolicies.filter(
-        (e) => e.permission === permission
-      );
-      // Entity Policies are "grants" in that only one needs to pass
-      // but we still want each check returned so we can see why they
-      // got access or got denied
-      const entityChecks = entityPermissionPolicies.map((policy) => {
-        return checkEntityPolicy(policy, context);
-      });
-      // Process them to see if any grant access
-      const grantedCheck = entityChecks.find((e) => e.response === "granted");
-      // If we did not find a check that grants access, AND we've passed
-      // all the system checks, then we set the response to "not-granted"
-      // and set the access to false
-      if (
-        entityChecks.length &&
-        !grantedCheck &&
-        response.response === "granted"
-      ) {
-        response.access = false;
-        response.response = "not-granted";
-      }
-      // Merge in the entity checks...
-      response.checks = [...response.checks, ...entityChecks];
+  // Is this policy configurable by the entity?
+  if (systemPolicy.entityConfigurable) {
+    // Has the entity provided a flag value?
+    if (entity?.features?.hasOwnProperty(permission)) {
+      flagging.hasFlag = true;
+      flagging.value = entity.features[permission];
+      flagging.type = "entity";
     }
   }
 
-  // log denied access information
+  // Feature Flags
+  // Passed in from the application when context is created,
+  // these override entity flags, so they are checked after
+  if (context.featureFlags.hasOwnProperty(permission)) {
+    flagging.hasFlag = true;
+    flagging.value = context.featureFlags[permission];
+    flagging.type = "feature";
+  }
+
+  // in all cases, if flag exists and false, access is denied
+  if (flagging.hasFlag && !flagging.value) {
+    const respValue = `disabled-by-${flagging.type}-flag` as PolicyResponse;
+    const disabledByFlagResponse = {
+      policy: permission,
+      access: false,
+      response: respValue,
+      code: getPolicyResponseCode(respValue),
+      checks: [],
+    } as IPermissionAccessResponse;
+    logResponse(disabledByFlagResponse, label);
+    return disabledByFlagResponse;
+  }
+
+  // required checks - things feature flags can not override
+  const requiredChecks: PermissionCheckFunction[] = [
+    checkParents,
+    checkServiceStatus,
+    checkAuthentication,
+    checkPrivileges,
+    checkOwner,
+    checkEdit,
+    checkLicense,
+    checkAssertions,
+  ];
+
+  // checks that feature flags can override
+  const overridableChecks: PermissionCheckFunction[] = [
+    checkAvailability,
+    checkEnvironment,
+  ];
+
+  let checkFns: PermissionCheckFunction[] = [];
+  // If there is a "feature" flag, set to true, for the policy
+  // we just run the required checks
+  if (flagging.hasFlag && flagging.value && flagging.type === "feature") {
+    checkFns = [...requiredChecks];
+  } else {
+    // otherwise we run all the checks
+    checkFns = [...overridableChecks, ...requiredChecks];
+  }
+
+  // execute the checks
+  const checks = checkFns.reduce((acc: IPolicyCheck[], fn) => {
+    acc = [...acc, ...fn(systemPolicy, context, entity)];
+    return acc;
+  }, []);
+
+  // Default to granted
+  const response: IPermissionAccessResponse = {
+    policy: permission,
+    access: true,
+    response: "granted",
+    code: getPolicyResponseCode("granted"),
+    checks: [],
+  };
+
+  // For system policies, all conditions must be met, so we can
+  // iterate through the checks and set the response to the first failure
+  // while still returning all the checks for observability
+  checks.forEach((check) => {
+    if (check.response !== "granted" && response.response === "granted") {
+      response.response = check.response;
+      response.code = check.code;
+      response.access = false;
+    }
+  });
+
+  response.checks = checks;
+
+  // Entity policies are treated as "grants" so we only need to pass one
+  if (entity) {
+    const entityPolicies: IEntityPermissionPolicy[] = getWithDefault(
+      entity,
+      "permissions",
+      []
+    );
+
+    const entityPermissionPolicies = entityPolicies.filter(
+      (e) => e.permission === permission
+    );
+    // Entity Policies are "grants" in that only one needs to pass
+    // but we still want each check returned so we can see why they
+    // got access or got denied
+    const entityChecks = entityPermissionPolicies.map((policy) => {
+      return checkEntityPolicy(policy, context);
+    });
+    // Process them to see if any grant access
+    const grantedCheck = entityChecks.find((e) => e.response === "granted");
+    // If we did not find a check that grants access, AND we've passed
+    // all the system checks, then we set the response to "not-granted"
+    // and set the access to false
+    if (
+      entityChecks.length &&
+      !grantedCheck &&
+      response.response === "granted"
+    ) {
+      response.access = false;
+      response.response = "not-granted";
+    }
+    // Merge in the entity checks...
+    response.checks = [...response.checks, ...entityChecks];
+  }
+
+  // log response
+  logResponse(response, label);
+
+  return response;
+}
+
+function logResponse(response: IPermissionAccessResponse, label: string): void {
   if (!response.access) {
     // tslint:disable-next-line:no-console
     console.info(
-      `checkPermission: ${label} ${permission} : ${response.response}`
+      `checkPermission: ${label} ${response.policy} : ${response.response}`
     );
     // tslint:disable-next-line:no-console
     console.dir(response);
     // tslint:disable-next-line:no-console
     console.info(`-----------------------------------------`);
   }
-  return response;
 }