+++ packages/common/test/content/computeProps.test.ts
@@ -3,41 +3,22 @@ import {
   deriveLocationFromItemExtent,
   getItemExtent,
 } from "../../src/content/_internal/computeProps";
-import { MOCK_AUTH } from "../mocks/mock-auth";
-import { ArcGISContextManager } from "../../src/ArcGISContextManager";
-import { IPortal } from "@esri/arcgis-rest-portal";
-import { IUser } from "@esri/arcgis-rest-types";
-import { IHubEditableContent, IModel } from "../../src";
+import { IHubEditableContent } from "../../src/core/types/IHubEditableContent";
+import { IHubRequestOptions, IModel } from "../../src/types";
+import { cloneObject } from "../../src/util";
+import { MOCK_HUB_REQOPTS } from "../mocks/mock-auth";
 
 describe("content computeProps", () => {
-  let authdCtxMgr: ArcGISContextManager;
+  let requestOptions: IHubRequestOptions;
   beforeEach(async () => {
-    // When we pass in all this information, the context
-    // manager will not try to fetch anything, so no need
-    // to mock those calls
-    authdCtxMgr = await ArcGISContextManager.create({
-      authentication: MOCK_AUTH,
-      currentUser: {
-        username: "casey",
-        privileges: ["portal:user:createItem"],
-      } as unknown as IUser,
-      portal: {
-        name: "DC R&D Center",
-        id: "BRXFAKE",
-        urlKey: "fake-org",
-        properties: {
-          hub: {
-            enabled: true,
-          },
-        },
-      } as unknown as IPortal,
-      portalUrl: "https://org.maps.arcgis.com",
-    });
+    requestOptions = cloneObject(MOCK_HUB_REQOPTS);
   });
 
   it("computeProps model boundary undefined", () => {
     const model: IModel = {
       item: {
+        type: "Feature Service",
+        id: "9001",
         created: new Date().getTime(),
         modified: new Date().getTime(),
         properties: {
@@ -48,14 +29,12 @@ describe("content computeProps", () => {
       // no boundary set
     } as IModel;
     const content: Partial<IHubEditableContent> = {
+      type: "Feature Service",
+      id: "9001",
       // no location set
     };
 
-    const chk = computeProps(
-      model,
-      content,
-      authdCtxMgr.context.requestOptions
-    );
+    const chk = computeProps(model, content, requestOptions);
 
     expect(chk.location?.type).toBe("custom");
   });
@@ -63,6 +42,8 @@ describe("content computeProps", () => {
   it("computeProps boundary defined as none", () => {
     const model: IModel = {
       item: {
+        type: "Feature Service",
+        id: "9001",
         created: new Date().getTime(),
         modified: new Date().getTime(),
         properties: {
@@ -73,16 +54,65 @@ describe("content computeProps", () => {
       // no boundary set
     } as IModel;
     const content: Partial<IHubEditableContent> = {
+      type: "Feature Service",
+      id: "9001",
       // no location set
     };
 
-    const chk = computeProps(
-      model,
-      content,
-      authdCtxMgr.context.requestOptions
-    );
+    const chk = computeProps(model, content, requestOptions);
 
     expect(chk.location?.type).toBe("none");
+  });
+
+  it("adds relative link when no slug is available", () => {
+    const model: IModel = {
+      item: {
+        type: "Feature Service",
+        id: "9001",
+        created: new Date().getTime(),
+        modified: new Date().getTime(),
+        properties: {
+          boundary: "none",
+        },
+      },
+      data: {},
+      // no boundary set
+    } as IModel;
+    const content: Partial<IHubEditableContent> = {
+      type: "Feature Service",
+      id: "9001",
+      // no location set
+    };
+
+    const chk = computeProps(model, content, requestOptions);
+
+    expect(chk.links.siteRelative).toBe("/maps/9001");
+  });
+
+  it("adds relative link when a slug is available", () => {
+    const model: IModel = {
+      item: {
+        type: "Feature Service",
+        id: "9001",
+        created: new Date().getTime(),
+        modified: new Date().getTime(),
+        properties: {
+          boundary: "none",
+        },
+      },
+      data: {},
+      // no boundary set
+    } as IModel;
+    const content: Partial<IHubEditableContent> = {
+      type: "Feature Service",
+      id: "9001",
+      slug: "my-slug",
+      // no location set
+    };
+
+    const chk = computeProps(model, content, requestOptions);
+
+    expect(chk.links.siteRelative).toBe("/maps/my-slug");
   });
 });