+++ packages/common/test/groups/HubGroups.test.ts
@@ -5,7 +5,6 @@ import {
   cloneObject,
   enrichGroupSearchResult,
   IHubRequestOptions,
-  setProp,
 } from "../../src";
 import * as HubGroupsModule from "../../src/groups/HubGroups";
 import * as FetchEnrichments from "../../src/groups/_internal/enrichments";
@@ -18,10 +17,10 @@ const TEST_GROUP: IGroup = {
   isInvitationOnly: false,
   owner: "dev_pre_hub_admin",
   description: "dev followers Content summary",
-  snippet: null,
+  snippet: undefined,
   tags: ["Hub Initiative Group", "Open Data"],
   typeKeywords: [],
-  phone: null,
+  phone: undefined,
   sortField: "title",
   sortOrder: "asc",
   isViewOnly: false,
@@ -104,11 +103,11 @@ describe("HubGroups Module:", () => {
       expect(chk.updatedDate).toEqual(new Date(GRP.modified));
       expect(chk.updatedDateSource).toEqual("group.modified");
       expect(chk.family).toEqual("team");
-      expect(chk.links.self).toEqual(
+      expect(chk.links?.self).toEqual(
         `https://some-server.com/gis/home/group.html?id=${GRP.id}`
       );
-      expect(chk.links.siteRelative).toEqual(`/teams/${GRP.id}`);
-      expect(chk.links.thumbnail).toEqual(
+      expect(chk.links?.siteRelative).toEqual(`/teams/${GRP.id}`);
+      expect(chk.links?.thumbnail).toEqual(
         `${hubRo.portal}/community/groups/${GRP.id}/info/${GRP.thumbnail}`
       );
       // Group Specific Props